리눅스 시간 기준 한구으로 변경

+++ backend/src/main/java/com/example/hackathon_server/bookconnected/entity/Library.java
@@ -16,7 +16,7 @@ import java.sql.Time;
 public class Library {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long libraryId;
+    private Integer libraryId;
 
     @Column
     private String libraryName;
